% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{args}
\alias{args}
\alias{findroot_args}
\alias{optimize_args}
\alias{integrate_args}
\alias{cg_args}
\alias{bfgs_args}
\alias{lbfgsb_args}
\alias{neldermead_args}
\alias{nlm_args}
\alias{richardson_args}
\title{Arguments}
\usage{
findroot_args()

optimize_args()

integrate_args()

cg_args()

bfgs_args()

lbfgsb_args()

neldermead_args()

nlm_args()

richardson_args()
}
\description{
Get an arguments list for internal methods with defaults set. These can
be adjusted and passed to their respective function.
}
