% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{fntl-package}
\alias{fntl}
\alias{fntl-package}
\title{fntl}
\description{
Package documentation
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewraim/fntl}
}

}
\author{
\strong{Maintainer}: Andrew M. Raim \email{andrew.raim@gmail.com} (\href{https://orcid.org/0000-0002-4440-2330}{ORCID})

Authors:
\itemize{
  \item Andrew M. Raim \email{andrew.raim@gmail.com} (\href{https://orcid.org/0000-0002-4440-2330}{ORCID})
}

}
