% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{univariate-optimization}
\alias{univariate-optimization}
\alias{goldensection}
\alias{optimize_brent}
\title{Univariate Optimization}
\usage{
goldensection(f, lower, upper, args)

optimize_brent(f, lower, upper, args)
}
\arguments{
\item{f}{Function to optimize.}

\item{lower}{Lower limit of search interval. Must be finite.}

\item{upper}{Upper limit of search interval. Must be finite.}

\item{args}{List of additional arguments from the function \code{optimize_args}.}
}
\value{
A list with the form of a \code{optimize_result} described in section
"Univariate Optimization" of the package vignette.
}
\description{
Univariate Optimization
}
\examples{
f = function(x) { x^2 - 1 }
args = optimize_args()
goldensection(f, 0, 10, args)
optimize_brent(f, 0, 10, args)

}
